Work in progress - no longer use sip2.linphone.org as outbound proxy. This is causing issues. Indeed, many tester are using sip.example.org and REGISTER to this domain. However, the flexisip server advertises itself as "sip2.linphone.org" in Record-Routes. This causes ACK to be delayed, because a DNS resolution is requested and a new TCP/TLS connection will be made for it. This is unefficient for nothing (the use case of having a proxy that doesn't advertises itself the same way as the client knows them is unusual). In practice, this problem was causing the failure of many tests in "Call recovery" suite, because these tests were simulating network disconnections just after call establishment. Due to the time needed for the ACK to be send, the ACK was never sent because network was cut before.

+++ src/core/core-chat-room.cpp
@@ -184,12 +184,13 @@ void CorePrivate::insertChatRoom (const shared_ptr<AbstractChatRoom> &chatRoom) 
 
 void CorePrivate::insertChatRoomWithDb (const shared_ptr<AbstractChatRoom> &chatRoom, unsigned int notifyId) {
 	L_ASSERT(chatRoom->getState() == ChatRoom::State::Created);
-	mainDb->insertChatRoom(chatRoom, notifyId);
+	if (mainDb->isInitialized()) mainDb->insertChatRoom(chatRoom, notifyId);
 }
 
 void CorePrivate::loadChatRooms () {
 	chatRooms.clear();
 	chatRoomsById.clear();
+	if (!mainDb->isInitialized()) return;
 	for (auto &chatRoom : mainDb->getChatRooms()) {
 		insertChatRoom(chatRoom);
 		chatRoom->getPrivate()->sendDeliveryNotifications();
@@ -351,7 +352,7 @@ void Core::deleteChatRoom (const shared_ptr<const AbstractChatRoom> &chatRoom) {
 		L_ASSERT(chatRoomsIt != d->chatRooms.end());
 		d->chatRooms.erase(chatRoomsIt);
 		d->chatRoomsById.erase(chatRoomsByIdIt);
-		d->mainDb->deleteChatRoom(conferenceId);
+		if (d->mainDb->isInitialized()) d->mainDb->deleteChatRoom(conferenceId);
 	} else
 		L_ASSERT(find(d->chatRooms, chatRoom) == d->chatRooms.end());
 }
